requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
txtai==7.1.0
python-dotenv==1.0.0
networkx==3.2.1
sentence-transformers==2.3.1